pyproject.toml
[tool.poetry]
name = "adt"
version = "0.1.0"
description = "An attempt to simplify ADT definition in Python."
authors = ["Alexandru Dinu <alex.dinu07@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "adt"}]

[tool.poetry.dependencies]
python = "^3.11"
lark-parser = ">=0.12.0,<0.13.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
